Extract Redux models
