+++ src/features/models.ts
@@ -1,0 +1,38 @@
+import { GridLoc } from '../components/board/grid';
+import { Direction } from '../components/room/Room';
+
+export enum PlayerColor {
+  WHITE = 'white',
+  BLUE = 'blue',
+  GREEN = 'green',
+  YELLOW = 'yellow',
+  RED = 'red',
+  PURPLE = 'purple',
+}
+
+export interface Player {
+  color: PlayerColor;
+}
+
+export interface Room {
+  name: string;
+  loc: GridLoc;
+  doorDirections: Direction[];
+  players: Player[];
+}
+
+export enum Floor {
+  BASEMENT,
+  GROUND,
+  UPPER,
+  ROOF,
+}
+
+export interface StackRoom {
+  possibleFloors: Floor[];
+}
+
+export interface FlippedRoom {
+  name: string;
+  doorDirections: Direction[];
+}
